+++ rollup-common.js
@@ -29,6 +29,7 @@ const PACKAGE_CLASS_PREFIXES = {
   '@lit-labs/ssr-client': '_$I',
   '@lit-labs/task': '_$J',
   '@lit-labs/router': '_$K',
+  '@lit-labs/observers': '_$L',
 };
 
 // Validate prefix uniqueness
@@ -408,12 +409,13 @@ export function litProdConfig({
             ]),
       ],
     },
-    ...bundled.map(({file, output, name}) =>
+    ...bundled.map(({file, output, name, format}) =>
       litMonoBundleConfig({
         file,
         output,
         name,
         terserOptions,
+        format,
       })
     ),
   ];
@@ -424,12 +426,13 @@ const litMonoBundleConfig = ({
   output,
   name,
   terserOptions,
+  format = 'umd',
   // eslint-disable-next-line no-undef
 } = options) => ({
   input: `development/${file}.js`,
   output: {
     file: `${output || file}.js`,
-    format: 'umd',
+    format,
     name,
     sourcemap: !CHECKSIZE,
   },